+++ src/wai/annotations/imgaug/isp/sub_images/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for the sub-images ISP.
+"""